Removed comments, redundant test code, and renamed activity.
